storage: decouple FormatForm from PartitionForm
When formatting an entire disk, we used to lean on PartitionForm because the
partition form also offers to format the newly created partition(-like).

That said, some fields from the partition form are not relevant when doing only
formatting:

 * size (of the partition or LVM LV)
 * name (of the LVM LV)

We now define another form that is strictly for formatting and use inheritance
to have the PartitionForm be a subclass of FormatForm.
