+++ subiquity/ui/views/filesystem/tests/test_partition.py
@@ -7,9 +7,7 @@ from subiquity.client.controllers.filesystem import FilesystemController
 from subiquity.common.filesystem import gaps
 from subiquity.models.filesystem import MiB, dehumanize_size
 from subiquity.models.tests.test_filesystem import make_model_and_disk
-from subiquity.ui.mount import common_mountpoints, suitable_mountpoints_for_existing_fs
 from subiquity.ui.views.filesystem.partition import (
-    FormatEntireStretchy,
     PartitionForm,
     PartitionStretchy,
 )
@@ -29,17 +27,6 @@ def make_partition_view(model, disk, **kw):
     return base_view, stretchy
 
 
-def make_format_entire_view(model, disk):
-    controller = mock.create_autospec(spec=FilesystemController)
-    base_view = BaseView(urwid.Text(""))
-    base_view.model = model
-    base_view.controller = controller
-    base_view.refresh_model_inputs = lambda: None
-    stretchy = FormatEntireStretchy(base_view, disk)
-    base_view.show_stretchy_overlay(stretchy)
-    return base_view, stretchy
-
-
 class PartitionViewTests(unittest.TestCase):
     def test_initial_focus(self):
         model, disk = make_model_and_disk()
@@ -51,6 +38,29 @@ class PartitionViewTests(unittest.TestCase):
                 return
         else:
             self.fail("Size widget not focus")
+
+    def test_create_partition_unaligned_size(self):
+        # In LP: #2013201, the user would type in 1.1G and the partition
+        # created would not be aligned to a MiB boundary.
+        unaligned_data = {
+            "size": "1.1G",  # Corresponds to 1181116006.4 bytes (not an int)
+            "fstype": "ext4",
+        }
+        valid_data = {
+            "mount": "/",
+            "size": 1127 * MiB,  # ~1.10058 GiB
+            "use_swap": False,
+            "fstype": "ext4",
+        }
+        model, disk = make_model_and_disk()
+        gap = gaps.Gap(device=disk, offset=1 << 20, size=99 << 30)
+        view, stretchy = make_partition_view(model, disk, gap=gap)
+        view_helpers.enter_data(stretchy.form, unaligned_data)
+        stretchy.form.size.widget.lost_focus()
+        view_helpers.click(stretchy.form.done_btn.base_widget)
+        view.controller.partition_disk_handler.assert_called_once_with(
+            stretchy.disk, valid_data, partition=None, gap=gap
+        )
 
     def test_create_partition(self):
         valid_data = {
@@ -240,37 +250,6 @@ class PartitionViewTests(unittest.TestCase):
             stretchy.disk, expected_data, partition=stretchy.partition, gap=None
         )
 
-    def test_format_entire_unusual_filesystem(self):
-        model, disk = make_model_and_disk()
-        fs = model.add_filesystem(disk, "ntfs")
-        fs.preserve = True
-        model._orig_config = model._render_actions()
-        view, stretchy = make_format_entire_view(model, disk)
-        self.assertEqual(stretchy.form.fstype.value, None)
-
-    def test_create_partition_unaligned_size(self):
-        # In LP: #2013201, the user would type in 1.1G and the partition
-        # created would not be aligned to a MiB boundary.
-        unaligned_data = {
-            "size": "1.1G",  # Corresponds to 1181116006.4 bytes (not an int)
-            "fstype": "ext4",
-        }
-        valid_data = {
-            "mount": "/",
-            "size": 1127 * MiB,  # ~1.10058 GiB
-            "use_swap": False,
-            "fstype": "ext4",
-        }
-        model, disk = make_model_and_disk()
-        gap = gaps.Gap(device=disk, offset=1 << 20, size=99 << 30)
-        view, stretchy = make_partition_view(model, disk, gap=gap)
-        view_helpers.enter_data(stretchy.form, unaligned_data)
-        stretchy.form.size.widget.lost_focus()
-        view_helpers.click(stretchy.form.done_btn.base_widget)
-        view.controller.partition_disk_handler.assert_called_once_with(
-            stretchy.disk, valid_data, partition=None, gap=gap
-        )
-
 
 class TestPartitionForm(unittest.TestCase):
     def _make_name_form(self, lvm_names, name_value):
@@ -279,32 +258,10 @@ class TestPartitionForm(unittest.TestCase):
         form.name.value = name_value
         return form
 
-    def _make_mount_form(
-        self,
-        mount_value,
-        *,
-        mountpoints=None,
-        existing_fs_type=None,
-        fstype_value="ext4",
-        remote_storage=False,
-    ):
-        form = mock.MagicMock(spec=PartitionForm)
-        form.mount.value = mount_value
-        form.mountpoints = mountpoints if mountpoints is not None else {}
-        form.existing_fs_type = existing_fs_type
-        form.fstype.value = fstype_value
-        form.remote_storage = remote_storage
-        return form
-
     def _make_size_form(self, size_str, max_size):
         form = mock.MagicMock(spec=PartitionForm)
         form.size_str = size_str
         form.max_size = max_size
-        return form
-
-    def _make_clean_mount_form(self, is_mounted):
-        form = mock.MagicMock()
-        form.model.is_mounted_filesystem.return_value = is_mounted
         return form
 
     def test_validate_name__no_lvm(self):
@@ -342,61 +299,6 @@ class TestPartitionForm(unittest.TestCase):
         form = self._make_name_form({"other-lv"}, "my-lv")
         self.assertIsNone(PartitionForm.validate_name(form))
 
-    def test_validate_mount__none(self):
-        # None mount value means "leave unmounted"; always accepted.
-        form = self._make_mount_form(None)
-        self.assertIsNone(PartitionForm.validate_mount(form))
-
-    def test_validate_mount__error_path_too_long(self):
-        form = self._make_mount_form("a" * 4096)
-        self.assertIsNotNone(PartitionForm.validate_mount(form))
-
-    def test_validate_mount__error_already_mounted(self):
-        form = self._make_mount_form(
-            "/home", mountpoints={"/home": mock.sentinel.device}
-        )
-        with mock.patch(
-            "subiquity.ui.views.filesystem.partition.labels.label",
-            return_value="sda",
-        ):
-            self.assertIsNotNone(PartitionForm.validate_mount(form))
-
-    def test_validate_mount__valid(self):
-        form = self._make_mount_form("/home")
-        self.assertIsNone(PartitionForm.validate_mount(form))
-
-    def test_validate_mount__warning_existing_fs_unsuitable_mountpoint(self):
-        # Mounting an existing filesystem at a "bad" common mountpoint shows a
-        # warning but is not an error.
-        unsuitable = next(
-            m
-            for m in common_mountpoints
-            if m not in suitable_mountpoints_for_existing_fs
-        )
-        form = self._make_mount_form(
-            unsuitable,
-            existing_fs_type="ext4",
-            fstype_value=None,
-        )
-        self.assertIsNone(PartitionForm.validate_mount(form))
-        form.mount.show_extra.assert_called_once()
-
-    def test_validate_mount__no_warning_existing_fs_suitable_mountpoint(self):
-        # Mounting an existing filesystem at a suitable mountpoint shows no warning.
-        form = self._make_mount_form(
-            "/home",
-            existing_fs_type="ext4",
-            fstype_value=None,
-        )
-        self.assertIsNone(PartitionForm.validate_mount(form))
-        form.mount.show_extra.assert_not_called()
-
-    @parameterized.expand([["boot", "/boot"], ["boot_efi", "/boot/efi"]])
-    def test_validate_mount__warning_remote_storage(self, _name, mount):
-        form = self._make_mount_form(mount, remote_storage=True)
-        self.assertIsNone(PartitionForm.validate_mount(form))
-        form.mount.show_extra.assert_called_once()
-
     @parameterized.expand(
         [
             ["empty", ""],
@@ -419,11 +321,3 @@ class TestPartitionForm(unittest.TestCase):
         result = PartitionForm.clean_size(form, val)
         expected = dehumanize_size(val if val[-1].isalpha() else val + "G")
         self.assertEqual(expected, result)
-
-    def test_clean_mount__mounted_filesystem_returns_path(self):
-        form = self._make_clean_mount_form(is_mounted=True)
-        self.assertEqual("/home", PartitionForm.clean_mount(form, "/home"))
-
-    def test_clean_mount__unmounted_filesystem_returns_none(self):
-        form = self._make_clean_mount_form(is_mounted=False)
-        self.assertIsNone(PartitionForm.clean_mount(form, "/home"))
